Makefile/packit: Set version to single number only for releases
RPMs and releases should reflect if it's built from a
release tag or some other commit.

Expected behavior is, that if you build a tag like "v4"
the result is `image-builder-cli-4.tar.gz` otherwise
it should visualize that this is NOT a release e.g.
`image-builder-cli-4-12-g30166b0.tar.gz`